// src/shared/media-type.ts
export enum MediaType {
  app_xhtml_xml = 'application/xhtml+xml',
  text_html = 'text/html',
  rdf_xml = 'application/rdf+xml',
  json = 'application/json',
  json_ld = 'application/ld+json',
  trig = 'application/trig',
  turtle = 'text/turtle',
  n_triples = 'application/n-triples',
  n_quads = 'application/n-quads',
  n3 = 'text/n3',
}